Issues with React global initialization fixed
It results in removal of `mocha-jsdom` which was actually one factor
causing this issue. There was another approach to actually utilize
`mocha-jsdom` and its `rerequire` function. However, this is unnecessary
and although it very performant it is a bottleneck - so it does not make
sense to keep it. React tests do not pollute DOM tree as they occur on
detached nodes yet React uses globally available DOM object during its
initializing. Test cases which actually may need some document
manipulation should handle its own cleanup.

It also results in ES2015 imports in all tests which is nice on its own.