More stable and transparent modal pane
Previous approach creates unclear dependency between modal and client
component. It proved to be unnecessary, so it was replaced with simpler
and clearer approach - simply passing a React element.

Modal component has been split to Modal (mainly a structural wrapper for
modal markup) and Manager which handles open, close and some global
changes to DOM.

This simpler approach requires modal element to manage their state and
trigger re-renders. But this is not an issue.