Button-like anchors click events fixes
Firefox (and possibly Safari) did not trigger click event on anchor
elements which did not have hypertext reference (href). It triggered
focus event the first time element was clicked. Another click had to be
made to trigger click event.

The fix required to prevent default action on the click event. It was
actually possible to replace already used stop propagation and later in
bubble phase check if the event has been prevented in cases where it
matters. The approach is actually better after discussion so has been
applied in this commit.