Improved table rendering
Table components now omit children from comparison during
`shouldComponentUpdate` phase. This approach does not support dynamic
column specification, but it is not required at this point. Comparable
strategy is cell caching but it required way more logic and after
figuring out all the edge cases, the speed was comparable.

Some of the findings are summarized in the new PROBLEMS document.