Moving to Node.js naming convention
We did not follow AirBnb guideline to use the same file name as the main
class in the source file which we did agree we dislike. But we did not
follow any other common convention. Node.js uses snake-case (mostly;
e.g., `child_process` module in Node standard library) so we it too from
now.