Improved tabs maintain active in internal state
This is the default behavior when `tabChange` prop callback is not
specified. If it is, everything behaves as before.

Tabs also have a new test spec. It is not complete and covers only
wrapping Tabs component which is not required. New test spec also covers
new behavior with tab change.