Workflow table performance improvements
Table itself is relatively fast and it does not waste time rendering
what did not change when a workflow is activated. Which subjectively
seems as the main issue on the first sight.

The main issue at the moment is that activating certain workflow
triggers API requests which in turn change that workflow. This causes
changes in the whole table - most of which are
meaningless. Nevertheless, the main issue is actually store
organization.