Graph balancing improvements
Everything is centered around the heaviest and deepest path (branch)
which is always centered. This path is always centered. It may seem
weird for nodes with even number node below. However, it is easier to
solve than figuring out conflicts around positioning node with more than
node above. It the result is not that bad.

Some issues with graph travesral and other things are also fixed.