Dev/test server mocks API
It can provide fixtures or generated data. The API is nowhere near
complete and probably shouldn't be. It can simply provide previously
captured data for test cases.

As part of this effort, very basic type declarations were created in
JSDoc, React.PropTypes and JSON schema formats.

As a benefit, it can generate random data (in correct format based on
JSON schema), so in simple cases, the webapp can be developed without
Qorus.