More robust setup for integration/acceptance tests
Additional test scenarios showed additional holes in previous
setup. Particarly unpredictability of DOM changes therefore there is
additional level of confirmation that the DOM is ready. It actually
showed possible low severity bug around setting a document title. But
for now this works as expected.

For improved level of tolerance, test instrumentation in the webapp was
required. This is only at the topmost component. Hopefully, when
stability of DOM changes improves, it could be removed.