UAT @wip tag replaced with @no-impl
Work in progress tag is great while working on a feature but bad to
commit since others use while work and it will break their
flow. Something else should be used for tests that have no
implementation yet but are worth commiting - it is @no-impl. It can be
used on CI server later.