Test server forking removed
Test server is conveniently started only during `test` and `test-ci` npm
scripts. Otherwise, it is up to the developer to start the server. This
will changed shortly but this is just an iteration after a lot try/fail
approaches.

Becase the test server is started and stopped by pre- and post-scripts
respectively, a PID file is used to maintain reference to the server
process. No other IPC mechanism is used at the moment (previous
Node-native IPC from child_process module has been removed).